[ci skip] Refactor includes to follow <boost/gil/...hpp> convention
This is non-functional refactoring of headers inclusion.

Group includes in blocks sorted alphabetically within block:
 - <boost/gil...hpp>
 - <boost/...hpp>
 - standard headers
 - anything else
Rename gil_concept.hpp to concept.hpp
Rename gil_config.hpp to config.hpp
Rename utilities.hpp to utility.hpp
Fix include guards to BOOST_GIL_*_HPP convention.
