+++ include/boost/gil/bit_aligned_pixel_iterator.hpp
@@ -10,8 +10,8 @@
 
 /*************************************************************************************************/
 
-#ifndef GIL_BIT_ALIGNED_PIXEL_ITERATOR_HPP
-#define GIL_BIT_ALIGNED_PIXEL_ITERATOR_HPP
+#ifndef BOOST_GIL_BIT_ALIGNED_PIXEL_ITERATOR_HPP
+#define BOOST_GIL_BIT_ALIGNED_PIXEL_ITERATOR_HPP
 
 ////////////////////////////////////////////////////////////////////////////////////////
 /// \file               
@@ -22,14 +22,15 @@
 ///
 ////////////////////////////////////////////////////////////////////////////////////////
 
-#include <functional>
+#include <boost/gil/config.hpp>
+#include <boost/gil/bit_aligned_pixel_iterator.hpp>
+#include <boost/gil/pixel_iterator.hpp>
 
-#include <boost/config.hpp>
 #include <boost/iterator/iterator_facade.hpp>
 
-#include "gil_config.hpp"
-#include "bit_aligned_pixel_reference.hpp"
-#include "pixel_iterator.hpp"
+#include <cstddef>
+#include <algorithm>
+#include <functional>
 
 namespace boost { namespace gil {
