+++ include/boost/gil/bit_aligned_pixel_reference.hpp
@@ -10,8 +10,8 @@
 
 /*************************************************************************************************/
 
-#ifndef GIL_BIT_ALIGNED_PIXEL_REFERENCE_HPP
-#define GIL_BIT_ALIGNED_PIXEL_REFERENCE_HPP
+#ifndef BOOST_GIL_BIT_ALIGNED_PIXEL_REFERENCE_HPP
+#define BOOST_GIL_BIT_ALIGNED_PIXEL_REFERENCE_HPP
 
 ////////////////////////////////////////////////////////////////////////////////////////
 /// \file               
@@ -22,9 +22,10 @@
 ///
 ////////////////////////////////////////////////////////////////////////////////////////
 
-#include <functional>
+#include <boost/gil/config.hpp>
+#include <boost/gil/channel.hpp>
+#include <boost/gil/pixel.hpp>
 
-#include <boost/config.hpp>
 #include <boost/mpl/accumulate.hpp>
 #include <boost/mpl/at.hpp>
 #include <boost/mpl/bool.hpp>
@@ -33,9 +34,7 @@
 #include <boost/mpl/push_back.hpp>
 #include <boost/mpl/vector.hpp>
 
-#include "gil_config.hpp"
-#include "pixel.hpp"
-#include "channel.hpp"
+#include <functional>
 
 namespace boost { namespace gil {
