+++ include/boost/gil/device_n.hpp
@@ -9,8 +9,8 @@
 */
 /*************************************************************************************************/
 
-#ifndef GIL_DEVICE_N_H
-#define GIL_DEVICE_N_H
+#ifndef BOOST_GIL_DEVICE_N_HPP
+#define BOOST_GIL_DEVICE_N_HPP
 
 
 ////////////////////////////////////////////////////////////////////////////////////////
@@ -21,16 +21,15 @@
 /// \date 2005-2009 \n Last updated on February 20, 2009
 ////////////////////////////////////////////////////////////////////////////////////////
 
-#include <cstddef>
+#include <boost/gil/config.hpp>
+#include <boost/gil/metafunctions.hpp>
+#include <boost/gil/utility.hpp>
 
-#include <boost/type_traits.hpp>
-#include <boost/config.hpp>
 #include <boost/mpl/range_c.hpp>
 #include <boost/mpl/vector_c.hpp>
+#include <boost/type_traits.hpp>
 
-#include "gil_config.hpp"
-#include "utilities.hpp"
-#include "metafunctions.hpp"
+#include <cstddef>
 
 namespace boost { namespace gil {
