+++ include/boost/gil/iterator_from_2d.hpp
@@ -9,8 +9,8 @@
 */
 /*************************************************************************************************/
 
-#ifndef GIL_ITERATOR_FROM_2D_H
-#define GIL_ITERATOR_FROM_2D_H
+#ifndef BOOST_GIL_ITERATOR_FROM_2D_HPP
+#define BOOST_GIL_ITERATOR_FROM_2D_HPP
 
 ////////////////////////////////////////////////////////////////////////////////////////
 /// \file               
@@ -21,12 +21,14 @@
 ///
 ////////////////////////////////////////////////////////////////////////////////////////
 
-#include <cassert>
+#include <boost/gil/config.hpp>
+#include <boost/gil/concept.hpp>
+#include <boost/gil/locator.hpp>
+#include <boost/gil/pixel_iterator.hpp>
+
 #include <boost/iterator/iterator_facade.hpp>
-#include "gil_concept.hpp"
-#include "gil_config.hpp"
-#include "pixel_iterator.hpp"
-#include "locator.hpp"
+
+#include <cassert>
 
 namespace boost { namespace gil {
