+++ include/boost/gil/rgb.hpp
@@ -10,8 +10,8 @@
 
 /*************************************************************************************************/
 
-#ifndef GIL_RGB_H
-#define GIL_RGB_H
+#ifndef BOOST_GIL_RGB_HPP
+#define BOOST_GIL_RGB_HPP
 
 ////////////////////////////////////////////////////////////////////////////////////////
 /// \file
@@ -21,12 +21,14 @@
 /// \date 2005-2007 \n Last updated on October 10, 2007
 ////////////////////////////////////////////////////////////////////////////////////////
 
-#include <cstddef>
+#include <boost/gil/config.hpp>
+#include <boost/gil/metafunctions.hpp>
+#include <boost/gil/planar_pixel_iterator.hpp>
+
 #include <boost/mpl/range_c.hpp>
 #include <boost/mpl/vector_c.hpp>
-#include "gil_config.hpp"
-#include "metafunctions.hpp"
-#include "planar_pixel_iterator.hpp"
+
+#include <cstddef>
 
 namespace boost { namespace gil {
